RF - move regression module to fmristat
Regression module in statistics only used in fmristat code, and name was
confusing.  Although the code is generally useful for iteration over
results, move to fmristat for now, and rename to reflect the uses.
