+++ nipy/modalities/fmri/fmristat/delay.py
@@ -20,7 +20,6 @@ from nipy.algorithms.utils.matrices import pos_recipr, recipr0
 # FIXME: This is broken. Don't know how to fix it
 #from nipy.fixes.scipy.stats.models.contrast import Contrast, ContrastResults
 
-from nipy.algorithms.statistics.regression import TOutput 
 
 class Contrast(object):
     """ Empty boggus class to get the docs building.
