+++ nipy/modalities/fmri/fmristat/outputters.py
